explicit int casts

+++ Classes/KIFTestActor.m
@@ -60,7 +60,7 @@
     self = [super init];
     if (self) {
         _file = [file retain];
-        _line = line;
+        _line = (int)line;
         _delegate = delegate;
         _executionBlockTimeout = [[self class] defaultTimeout];
     }
